requirements.txt
et-xmlfile==1.1.0
grpcio==1.59.0
grpcio-tools==1.59.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
protobuf==4.24.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
